@@ -1,9 +1,9 @@
 package com.mm.luna.ui.nba;
 
 import android.content.Intent;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
 
 import com.chad.library.adapter.base.BaseQuickAdapter;
 import com.mm.luna.R;
