@@ -1,6 +1,6 @@
 package com.mm.luna.view.statusLayoutView;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.View;
 import android.view.ViewGroup;
 
